fix integers

+++ binary.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 )
 
-func ToBinary(n int) string {
+func ToBinary(n float64) string {
 	var s string
-	for i := n; i > 0; i /= 2 {
+	for i := int(n); i > 0; i /= 2 {
 		s += fmt.Sprint(i % 2)
 	}
 	Reverse(&s)
