+++ hexadecimal.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 )
 
-func ToHex(n int) string {
+func ToHex(n float64) string {
 	var s string
-	for i := n; i > 0; i /= 16 {
+	for i := int(n); i > 0; i /= 16 {
 		s += fmt.Sprint(i % 16)
 	}
 	Reverse(&s)
